feat: implement user signup and login pages with authentication

+++ backend/src/services/recipes.js
@@ -1,13 +1,10 @@
 import { Recipe } from '../db/models/recipe.js'
 
-export async function createRecipe({
-  title,
-  author,
-  ingredients,
-  image,
-  tags,
-}) {
-  const recipe = new Recipe({ title, author, ingredients, image, tags })
+export async function createRecipe(
+  userId,
+  { title, ingredients, image, tags },
+) {
+  const recipe = new Recipe({ title, author: userId, ingredients, image, tags })
   return await recipe.save()
 }
 
@@ -35,16 +32,17 @@ export async function getRecipeById(recipeId) {
 }
 
 export async function updateRecipe(
+  userId,
   recipeId,
-  { title, author, ingredients, image, tags },
+  { title, ingredients, image, tags },
 ) {
   return await Recipe.findOneAndUpdate(
-    { _id: recipeId },
-    { $set: { title, author, ingredients, image, tags } },
+    { _id: recipeId, author: userId },
+    { $set: { title, ingredients, image, tags } },
     { new: true },
   )
 }
 
-export async function deleteRecipe(recipeId) {
-  return await Recipe.deleteOne({ _id: recipeId })
+export async function deleteRecipe(userId, recipeId) {
+  return await Recipe.deleteOne({ _id: recipeId, author: userId })
 }
